refactor: Clean up ModelProperty code
