+++ openapi_python_client/parser/properties/__init__.py
@@ -9,7 +9,7 @@ from ..errors import PropertyError, ValidationError
 from ..reference import Reference
 from .converter import convert, convert_chain
 from .enum_property import EnumProperty
-from .model_property import ModelProperty
+from .model_property import ModelProperty, build_model_property
 from .property import Property
 from .schemas import Schemas
 
@@ -232,98 +232,6 @@ def _string_based_property(
             pattern=data.pattern,
             nullable=data.nullable,
         )
-
-
-def build_model_property(
-    *, data: oai.Schema, name: str, schemas: Schemas, required: bool, parent_name: Optional[str]
-) -> Tuple[Union[ModelProperty, PropertyError], Schemas]:
-    """
-    A single ModelProperty from its OAI data
-
-    Args:
-        data: Data of a single Schema
-        name: Name by which the schema is referenced, such as a model name.
-            Used to infer the type name if a `title` property is not available.
-        schemas: Existing Schemas which have already been processed (to check name conflicts)
-    """
-    required_set = set(data.required or [])
-    required_properties: List[Property] = []
-    optional_properties: List[Property] = []
-    relative_imports: Set[str] = set()
-
-    class_name = data.title or name
-    if parent_name:
-        class_name = f"{utils.pascal_case(parent_name)}{utils.pascal_case(class_name)}"
-    ref = Reference.from_ref(class_name)
-
-    all_props = data.properties or {}
-    for sub_prop in data.allOf or []:
-        if isinstance(sub_prop, oai.Reference):
-            source_name = Reference.from_ref(sub_prop.ref).class_name
-            sub_model = schemas.models.get(source_name)
-            if sub_model is None:
-                return PropertyError(f"Reference {sub_prop.ref} not found"), schemas
-            required_properties.extend(sub_model.required_properties)
-            optional_properties.extend(sub_model.optional_properties)
-            relative_imports.update(sub_model.relative_imports)
-        else:
-            all_props.update(sub_prop.properties or {})
-            required_set.update(sub_prop.required or [])
-
-    for key, value in all_props.items():
-        prop_required = key in required_set
-        prop, schemas = property_from_data(
-            name=key, required=prop_required, data=value, schemas=schemas, parent_name=class_name
-        )
-        if isinstance(prop, PropertyError):
-            return prop, schemas
-        if prop_required and not prop.nullable:
-            required_properties.append(prop)
-        else:
-            optional_properties.append(prop)
-        relative_imports.update(prop.get_imports(prefix=".."))
-
-    additional_properties: Union[bool, Property, PropertyError]
-    if data.additionalProperties is None:
-        additional_properties = True
-    elif isinstance(data.additionalProperties, bool):
-        additional_properties = data.additionalProperties
-    elif isinstance(data.additionalProperties, oai.Schema) and not any(data.additionalProperties.dict().values()):
-        # An empty schema
-        additional_properties = True
-    else:
-        assert isinstance(data.additionalProperties, (oai.Schema, oai.Reference))
-        additional_properties, schemas = property_from_data(
-            name="AdditionalProperty",
-            required=True,  # in the sense that if present in the dict will not be None
-            data=data.additionalProperties,
-            schemas=schemas,
-            parent_name=class_name,
-        )
-        if isinstance(additional_properties, PropertyError):
-            return additional_properties, schemas
-        relative_imports.update(additional_properties.get_imports(prefix=".."))
-
-    prop = ModelProperty(
-        reference=ref,
-        required_properties=required_properties,
-        optional_properties=optional_properties,
-        relative_imports=relative_imports,
-        description=data.description or "",
-        default=None,
-        nullable=data.nullable,
-        required=required,
-        name=name,
-        additional_properties=additional_properties,
-    )
-    if prop.reference.class_name in schemas.models:
-        error = PropertyError(
-            data=data, detail=f'Attempted to generate duplicate models with name "{prop.reference.class_name}"'
-        )
-        return error, schemas
-
-    schemas = attr.evolve(schemas, models={**schemas.models, prop.reference.class_name: prop})
-    return prop, schemas
 
 
 def build_enum_property(
